backend/DietAnalysisFunction/requirements.txt
azure-functions
pandas
azure-storage-blob
numpy
